net: move Endpoint to a new mod
